Create container when needed after user confirmation
When a container doesn't exist yet, and the user tries to operate on
said container, we now ask the user if we can create a stopped container
for them.  This easens the flow when trying to run scripts for a stack
that hasn't fully booted (yet).

Pairing with @erikap.

+++ mu
@@ -46,6 +46,30 @@ function ensure_mu_cli_docker() {
     fi
 }
 
+####
+## Ensure the container exists or ask the user if we can create it?
+##
+## Uses first argument as the service name in $service
+####
+function confirm_existing_service() {
+    service=$1
+    available_container_id=`docker-compose $(print_source_docker_files) ps -q $service`
+    if [[ $? -ne 0 ]]; then
+        echo "Service $service not found"
+        return 1;
+    elif [[ -z $available_container_id ]]; then
+        # ask user if we can create a stopped container
+        echo "Container for $service does not exist yet.  May we create a container for $service without starting it?"
+        read -p "Create container? [Y/n]: " -n 1 -s -r INPUT
+        case ${INPUT:-Y} in
+            [Yy]*) `docker-compose $(print_source_docker_files) up --no-start $service` ;;
+            [Nn]*) echo "NOT creating container"; return 1 ;;
+        esac
+    else
+        return 0
+    fi
+}
+
 function print_commands_documentation() {
     service=$1
     command=$2
@@ -70,20 +94,27 @@ function print_source_docker_files() {
 
 function print_service_documentation() {
     service=$1
-    available_container_id=`docker-compose $(print_source_docker_files) ps -q $service`
-    mkdir -p /tmp/mu/cache/$available_container_id
-    docker cp $available_container_id:/app/scripts /tmp/mu/cache/$available_container_id 2> /dev/null
-    local_cat_command="cat /tmp/mu/cache/$available_container_id/scripts/config.json"
-    if test -f "/tmp/mu/cache/$available_container_id/scripts/config.json"; then
-        supported_commands=`sh -c "$interactive_cli bash -c \"$local_cat_command | $jq_documentation_filter_commands\""`
-        for supported_command in $supported_commands
-        do
-            print_commands_documentation $service $supported_command
-            print_text_block "" ""
-        done
-        echo ""
+    if confirm_existing_service $service
+    then
+        service=$1
+        available_container_id=`docker-compose $(print_source_docker_files) ps -q $service`
+        mkdir -p /tmp/mu/cache/$available_container_id
+        docker cp $available_container_id:/app/scripts /tmp/mu/cache/$available_container_id 2> /dev/null
+        local_cat_command="cat /tmp/mu/cache/$available_container_id/scripts/config.json"
+        if test -f "/tmp/mu/cache/$available_container_id/scripts/config.json"; then
+            supported_commands=`sh -c "$interactive_cli bash -c \"$local_cat_command | $jq_documentation_filter_commands\""`
+            for supported_command in $supported_commands
+            do
+                print_commands_documentation $service $supported_command
+                print_text_block "" ""
+            done
+            echo ""
+        else
+            print_text_block "  no scripts found" \
+                             ""
+        fi
     else
-        print_text_block "  no scripts found" \
+        print_text_block "Cannot provide docs for $service without a container" \
                          ""
     fi
 }
@@ -324,13 +355,22 @@ then
             exit 0
         fi
 
+        # get hold of the container that defines the script
+        confirm_existing_service $service
+        service_exists=$?
         container_id=`docker-compose $(print_source_docker_files) ps -q $service`
-        if [[ -z $container_id ]] ;
+        if [[ $? -ne 0 ]] ; # the service doesn't exist in the docker-compose
         then
             echo ""
             print_available_services_information
             exit 1
         fi
+        if [[ $service_exists -ne 0 ]]
+        then
+            echo "Cannot execute script without existing container"
+            exit 1 # Service did not exist and user refused to create
+        fi
+
         mkdir -p /tmp/mu/cache/$container_id
         docker cp $container_id:/app/scripts /tmp/mu/cache/$container_id
         cat_command="cat /tmp/mu/cache/$container_id/scripts/config.json"
